checkpoint 50 assignment technically complete
