+++ Blocly/app/src/main/java/io/bloc/android/blocly/api/network/NetworkRequest.java
@@ -1,11 +1,7 @@
 package io.bloc.android.blocly.api.network;
 
-import android.util.Log;
-
-import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.net.MalformedURLException;
 import java.net.URL;
 
@@ -49,36 +45,4 @@ public abstract class NetworkRequest<Result> {
         return inputStream;
     }
 
-    public static class GetFeedsNetworkRequest extends NetworkRequest {
-
-        String[] feedUrls;
-
-        public GetFeedsNetworkRequest(String... feedUrls) {
-            this.feedUrls = feedUrls;
-        }
-
-        @Override
-        public Object performRequest() {
-            for (String feedUrlString : feedUrls) {
-                InputStream inputStream = openStream(feedUrlString);
-                if (inputStream == null) {
-                    return null;
-                }
-                try {
-                    BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(inputStream));
-                    String line = bufferedReader.readLine();
-                    while (line != null) {
-                        Log.v(getClass().getSimpleName(), "Line: " + line);
-                        line = bufferedReader.readLine();
-                    }
-                    bufferedReader.close();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                    setErrorCode(ERROR_IO);
-                    return null;
-                }
-            }
-            return null;
-        }
-    }
 }
